stocks/requirements.txt
requests
finnhub-python
